!XB (Cry3DEngine) fixed compiler error in regards to signed unsigned comparison
Copied from Perforce
 Change: 1873334

+++ Code/CryEngine/Cry3DEngine/RenderMeshMerger.cpp
@@ -252,7 +252,7 @@ void CRenderMeshMerger::MakeListOfAllCRenderChunks(SMergeInfo& info)
 		CRenderChunk Ch = newMatInfo;
 		for (uint32 i = Ch.nFirstIndexId; i < Ch.nFirstIndexId + Ch.nNumIndices; i++)
 		{
-			assert(i >= 0 && i < pRM->GetIndicesCount());
+			assert(i >= 0 && i < static_cast<uint32>(pRM->GetIndicesCount()));
 			assert(pSrcInds[i] >= Ch.nFirstVertId && pSrcInds[i] < Ch.nFirstVertId + Ch.nNumVerts);
 			assert((int)pSrcInds[i] < pRM->GetVerticesCount());
 		}
